src/requirements.txt
agate==1.7.0
astroid==2.15.6
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dbt-core==1.5.3
dbt-extractor==0.4.1
dbt-postgres==1.5.3
dill==0.3.6
distlib==0.3.6
filelock==3.12.2
Flask==2.3.2
future==0.18.3
greenlet==2.0.2
hologram==0.0.16
identify==2.5.24
idna==3.4
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.1.3
mashumaro==3.6
mccabe==0.7.0
minimal-snowplow-tracker==0.0.2
msgpack==1.0.5
networkx==2.8.8
nodeenv==1.8.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parsedatetime==2.4
pathspec==0.11.1
platformdirs==3.8.1
pre-commit==3.3.3
protobuf==4.23.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pylint==2.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
PyYAML==6.0
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
termcolor==2.3.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
Werkzeug==2.3.6
wrapt==1.15.0
